requirements.txt
flask
flask-cors
opencv-python
numpy
tensorflow
pillow
pyzbar
gunicorn
requests
python-dotenv